move http stuff in package outside of jvm
