+++ openwebstart/src/main/java/com/openwebstart/http/ConnectionUtils.java
@@ -1,4 +1,4 @@
-package com.openwebstart.jvm.io;
+package com.openwebstart.http;
 
 import net.adoptopenjdk.icedteaweb.Assert;
