+++ openwebstart/src/main/java/com/openwebstart/http/HttpResponse.java
@@ -1,4 +1,4 @@
-package com.openwebstart.jvm.io;
+package com.openwebstart.http;
 
 
 import net.adoptopenjdk.icedteaweb.Assert;
@@ -7,7 +7,6 @@ import java.io.Closeable;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.HttpURLConnection;
-import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
