+++ openwebstart/src/main/java/com/openwebstart/http/Subscription.java
@@ -1,0 +1,10 @@
+package com.openwebstart.http;
+
+@FunctionalInterface
+public interface Subscription {
+
+    /**
+     * Unsusbscribe / unregister the handling that is defined by the {@link Subscription} instance.
+     */
+    void unsubscribe();
+}
