+++ openwebstart/src/main/java/com/openwebstart/jvm/LocalRuntimeManager.java
@@ -1,9 +1,9 @@
 package com.openwebstart.jvm;
 
+import com.openwebstart.http.DownloadInputStream;
+import com.openwebstart.http.HttpGetRequest;
+import com.openwebstart.http.HttpResponse;
 import com.openwebstart.jvm.func.Result;
-import com.openwebstart.jvm.io.DownloadInputStream;
-import com.openwebstart.jvm.io.HttpGetRequest;
-import com.openwebstart.jvm.io.HttpResponse;
 import com.openwebstart.jvm.json.CacheStore;
 import com.openwebstart.jvm.json.JsonHandler;
 import com.openwebstart.jvm.listener.Registration;
